text-reply-system/requirements.txt
PyYAML>=6.0
numpy>=1.26
tqdm>=4.66
mlx>=0.17.0
mlx-lm>=0.16.0
